Clean up abandoned transactions with rollback
In ODBC, an explicit rollback is needed to end an abandoned transaction.
This was previously done as part of the clean up for odbc_tx, but was
incorrectly removed to suppress additional exceptions being generated by
the destructor while unwinding as a result of some other error.  The
clean up has been restored as a "quiet" rollback, i.e. without
generating exceptions.  This change is part of addressing a bug that
appears to cause an abandoned transaction to run into the next one.