Change how index creation errors are reported
Index creation errors previously have been logged as warnings, but in
most cases they are expected and not a concern.  They are nearly
always resulting from trying to create a B-tree index on a column that
includes a large varchar (>10000), where such an index is probably not
very useful.  This commit changes the warning to an informational
notice.